+++ tests/test_main.py
@@ -1,5 +1,6 @@
 '''Test functions in __main__.py'''
 from pathlib import Path
+from unittest.mock import Mock
 import pytest
 import mock
 import validate
@@ -80,6 +81,17 @@ def test_path_writable_io_error(mock_os_access):
 
     with pytest.raises(IOError):
         validate.__main__.path_writable(test_path)
+
+@mock.patch('validate.validators.bam.pysam')
+def test_file_bam_empty(mock_pysam):
+    '''Tests Value error for valid BAM header with no reads'''
+    mock_alignment_file = Mock()
+    mock_alignment_file.head.return_value = iter([])
+    mock_pysam.AlignmentFile.return_value = mock_alignment_file
+    test_path = Path('empty/valid/bam')
+
+    with pytest.raises(ValueError):
+        validate.validators.bam.validate_bam_file(test_path)
 
 def generate_md5_for_file_success():
     '''Tests successful generate_md5'''