+++ validate/validators/bam.py
@@ -9,4 +9,8 @@ def validate_bam_file(path):
     except pysam.SamtoolsError as err:
         raise ValueError("samtools bam check failed. " + str(err)) from err
 
+    bam_head: pysam.IteratorRowHead = pysam.AlignmentFile(str(path)).head(1)
+    if next(bam_head, None) is None: #if the iterator is exhausted, next() returns None
+        raise ValueError("pysam bam check failed. No reads in " + str(path))
+
     return True